requirements.txt
tensorflow==2.11.0
opencv-python==4.8.0.76
mediapipe==0.10.0
numpy==1.23.5
Pillow==9.3.0
